Update tablespace creation command to allow PVC size to be set
Continuing with the previous sets of commits, this patch allows for a
user to specify the size of the a tablespace that is being created
alongside a PostgreSQL cluster.

To make the pgo client CLI still be clean, this also introduces changes
in terms of how one creates a tablespace. This moves the tablespace
flag to start using name/value parameters, e.g.

    pgo create cluster hacluster \
        --tablespace=name=ts:storageconfig=nfsstorage:pvcsize=10Gi

which makes it clearer which parameters need to be used. The "name" and
"storageconfig" parameters are required.

Issue: [ch7227]
Issue: #1080