Remove direct support for "pgo benchmark"
While a helpful feature for profiling PostgreSQL clusters deployed
by the Operator, "pgo benchmark" also added to the overall support
burden for the system, and does not reflect the full set of features
that one may want to utilize with "pgbench", which it is based on
(and derived from "crunchy-pgbench").

As "crunchy-pgbench" exists and allows for further customizations,
clusters deployed by the PostgreSQL Operator can be benchmarked,
but this is not going to be a feature that is supported.