Report errors in SQL policies at apply time
A change introduced into PostgreSQL 4.2 by yours truly accidentally
broke the ability to report back in real-time if there were SQL syntax
errors when attempting to add a policy to a managed PostgreSQL cluster.
This fixes the issue by reading the error from STDERR input and relaying
it back to the user.

Issue: #1266