Remove pg_basebackup functionality
While 4.2 removed access from performing pg_basebackup due to the
introduction of pgBackRest as the standard method for physical
backups, there was still many artifacts of the legacy pg_basebackup
functionality, including an extra CRD that was not being utilized
at all.

This commit removes the artificats in a minimally invasive way
(as some of the "code additions" were to minimize the effort of
the transition, though a very quick read would indicate that said
sections of code could be improved) and removes the pgbackups CRD.

This commit also inadvertently deprecates "pgo delete backup" as
it was not being used for the logical backups at all, but this
could be rectified in a future commit.

Issue: [ch7154]