Allow pod anti-affinity rules to be overriden for pgBackRest & pgBouncer
v4.2.0 introduced the ability to set pod anti-affinity rules for pods
deployed as part of a PostgreSQL cluster, which include the pgBackRest
repository and the optional pgBouncer deployment. However, there may be
cases where a user may not want these rules to be universally applied,
e.g. where PostgreSQL instances should be "required" to be on different
nodes, but pgBackRest only has "preferred" anti-affinity.

This patch introduces the ability to set the pod anti-affinity rules
specifically for pgBackRest and pgbouncer as well. This is exposed to
the end user with the `--pod-anti-affinity-pgbackrest` and
`--pod-anti-affinity-pgbouncer` on `pgo create cluster`, which set the
anti-affinity rules for pgBackRest and pgBouncer respectively.

This can also be set as part of the "pgo.yaml" configuration file, using
the `PodAntiAffinityPgBackRest` and `PodAntiAffinityPgBouncer` keys for
pgBackRest and pgBouncer respectively.

This commit also creates a change in the behavior of the "required"
pod anti-affinity setting. Deployments that leverage the "required" mode
of anti-affinity set their pods to look for other pods with the
"required" label in their group. This is to ensure that pods that have
the lesser "preferred" label don't necessarily cause a "required" pod
to fail to be scheduled to a node.

Issue: [ch7136]