Remaining Changes for Standby Cluster Support
This commit completes the changes required to fully implement support
for standby clusters.  This includes the ability to enable and disable
standby mode within a cluster.

Standby mode can be disabled within a PG cluster that currently acting
as a standby (i.e. replicating WAL from an S3 pgBackRest repo that
is being populated by another remote PG cluster) by changing the
standby parameter on the pgcluster CR from 'true' to 'false'.  This can
be done by directly modifying the 'standby parameter on pgcluster CR
itself, or by using the 'pgo update cluster' CLI command and/or its
associate API server endpoint. When using the CLI, a new
'--disable-standby' flag has been provided to ensure the proper
parameter is set when calling the cluster update endpoint to disable
standby mode.

Standby mode is specifically disabled by removing the 'standby_cluster'
configuration from the cluster's DCS, which in turn triggers the
promotion of the cluster from a standby server to a regular PostgreSQL
server (i.e. a server that can accept writes and is no longer
replicating from another DB).  Additionally, if pgbouncer is enabled in
the cluster, the it will be updated at this time to rotate the
pgbouncer passwords, and therefore enable pgbouncer for the cluster.
When disabling standby mode via the CLI a warning is shown to ensure the
database the standby is replicating from has first been properly
shutdown (e.g. to prevent both databases from accepting writes, and
therefore trying to write WAL to the same S3 pgBackRest repo).

In support of ensuring the proper shutdown of a cluster prior to
promoting a new standby, the scaledown command and associated API
server endpoint now allows the user to shutdown a PG cluster.  When
using the CLI (specifically command 'pgo scaledown'), this is done by
specifying the '--shutdown' flag. In order to use this flag and shutdown
a cluster, all replicas must first be scaled down and removed,
otherwise an error will be returned.  Assuming all replicas have been
removed, this flag will result in all remaining deployments for the
cluster, e.g. the primary PG DB deployment, the pgBackRest repo
deployment, and if enabled the pgbouncer deployment, being scaled to 0,
effectively terminating any active pods for those deployments and
shutting down the cluster.  This ultimately allows for a
clean/structured shutdown process to ensure a database is properly and
cleanly shutdown prior to promoting a standby cluster.

As described above, this commit also allows standby mode to be enabled
within a cluster (assuming it has the proper S3 configuration).  This
allows an existing cluster to be turned into a standby cluster after
the the standby that is replicating from it has been promoted.  Before
enabling standby mode, the cluster must first be fully shutdown (using
the commands/endpoints described above).  From there the 'pgo
update cluster' command can be utilized in conjunction with the new
'--enable-standby' flag to enable standby mode.  This will delete and
recreate any PVCs for the cluster, allowing the cluster to be
bootstrapped from scratch as a standby.  Since this is a destructive
action (i.e. it results in the deletion of PVCs and therefore
potentially all cluster data), a warning is shown when enabling standby
mode, which indicates that data will only be retained per the retention
policy set for any PV's and/or Storage Classes.

Once standby mode has been enabled, the cluster can then be started,
and it will then bootstrap as a standby.  This is now possible using
the '--startup' flag provided with the 'pgo scale' command (or by
calling the associated API endpoint directly with the proper request
parameters), which results in the deployments for the clusters primary
DB and all associated services (e.g. the pgBackRest repo, as well as
pgBouncer if enabled) being scaled back up to 1.  This will effectively
start the cluster again from scratch, only this time with the proper
standby configuration needed to bootstrap the cluster as a standby.

It should be noted that in support of the '--startup' flag, the
'--replica-count' flag no longer defaults to 1 when using the PGO CLI,
and now defaults to 0.  This allows the '--startup' flag to be
specified with a replica count of 0 to start the cluster without any
replicas.  If '--startup' is specified along with a replica count
greater than zero, then the cluster will first be started (assuming it
is currently shutdown) prior to creating any replicas.

And finally, a few additional defects were corrected when implementing
the above:
- The original primary can now be removed following a failover using
the 'pgo scaledown' command (or associated API server endpoint)
- It is now possible to delete a cluster that was not properly/full
initialized.  Specifically, a panic will no longer occur when
attempting to disable autofailover for a cluster who's DCS was never
created due to a failed cluster bootstrap.