add pgo-client package
The pgo-client package replaces the pgo/api package. The motivation
behind this modification is to provide a separate 'client' api for
interfacing with the pgo api server.

The package is broken down into the following components:

* api - provides an interface for all supported operations.
* v4/http - provides an http implementation for all supported operations
over HTTP.

The longer term vision of these changes is intended to facilitate
testing and making it easier for other client based applications to
interface with the api server.