Improve initial user creation with "pgo create cluster"
This commit adds several new features to the "pgo create cluster"
workflow:

- Introduce the "--username" / "-u" flag, to specify the user
account the user wishes to log in as
- Allow the user to specify "--password-length"
- Add the "--show-system-accounts" flag to show information about
the system accounts, i.e. username / passwords

Additionally, "pgo create cluster" now returns the username/password
information for the PostgreSQL users that are added at cluster creation
time. This is available in the API response as well.

There are some additional code refactors to reorganize some of the
logic for present and future changes.

Issue: [ch5416]