Trigger Replica Creation Post-Restore
As a result of a recent refactor of the job controller, replicas are no
longer being created following a restore.  This change updates the
Operator restore logic to ensure replicas are properly re-created
following a restore, specifically in a way that aligns with the new
structure for the job controller.