Allow user to specify PostgreSQL database name at cluster creation
A user can now specify the name of a database that is created during
cluster creation using the "--database" or "-d" flag on the
"pgo create cluster" command. The name of the database is returned as
part of the result.

Additionally, a few changes were made to how the initial database name
is selected. It now follows this heuristic:

1. If specified, used the value specified by "--database" or "-d"
2. If specified, use the value that is set in pgo.yaml
3. If nothing is specified, use the name of the cluster that is used for
"pgo create cluster"

To help make this change more user-friendly, the "default value" that
is supplied in the "pgo.yaml" file (i.e. "userdb") is now removed. This
aligns with user behavior and requests to user the name that is supplied
as part of "pgo create cluster"

Issue: [ch7506]