Rename ShowCluster* messages and API interfaces.
Using the term 'show' is a misnomer from the perspective of the API. The
API provides the functionality to allow a client to retrieve or 'get'
information about a cluster. The client is then responsible for the
presentation or 'showing' of that information.  Therefore, these changes
rename the request/response messages and the supporting API interface
functions to be more correct in what function they perform.