Enable TLS collections in PostgreSQL clusters
Though there were ways to configured TLS-enabled PostgreSQL via the
PostgreSQL Operator, it required a fair bit of work. Given this is a
common deployment requirement, this commit introduces the ability to
enable PostgreSQL clusters to use TLS with a few flags and with the
appropriate setup in your Kubernetes environment.

TLS allows secure TCP connections to PostgreSQL, and the PostgreSQL
Operator makes it easy to enable this PostgreSQL feature. The TLS
support in the PostgreSQL Operator does not make an opinion about your
PKI, but rather loads in your TLS key pair that you wish to use for the
PostgreSQL server as well as its corresponding certificate authority
(CA) certificate. Both of these Secrets are required to enable TLS
support for your PostgreSQL cluster when using the PostgreSQL Operator,
but it in turn allows seamless TLS support.

This commit introduces three new flags to the "pgo create cluster"
command, including

- `--server-tls-secret`: the name of the Secret that contains the TLS
key pair to be used for the PostgreSQL server
- `--server-ca-secret`: the name of the Secret that contains the CA
certificate that the PostgreSQL cluster will use

Both of these flags are required in order to enable TLS in a PostgreSQL
cluster -- in fact, if both are included, then TLS is automatically
enabled!

The Kubernetes Secrets containing the TLS key pair and the CA
certificate must be in the same namespace as the PostgreSQL Cluster that
is using them.

To create the Secrets, use the following steps:

1. "TLS Secret": create the Secret using `kubectl create secret tls`
2. "CA Secret": create the Secret using `kubectl create secret generic`
and set the key name to "ca.crt"

The additional flag provided in "pgo create cluster" is the "--tls-only"
flag, which will force the PostgreSQL cluster to only accept TLS
connections.

Issue: [ch7493]
Issue: #1180