Set FSGroup to postgres UID universally
The storage configuration specs of the PostgreSQL Operator allowed for
a parameter to configure the FSGroup in the Pod Security Context of
any Pod mounting PostgreSQL data directories to the postgres UID, which
for the purposes of the Operator ecosystem is 26. However, this is not
something that is configurable, but is in fact a constant: if the
postgres process is unable to access the data directory, then the
database will not work at all.

This moves the FSGroup to a constant that can be universally accessed
when the Pod Security Context is set up. Additionally, the underlying
code for generating the configuration for the Pod Security Context is
changed as well, to move more towards the modern patterns for applying
these contexts.

The side effect of this is that it is now easier to configure and
install the PostgreSQL Operator, which is always a nice bonus.