Add pg1-path to pgbackrest.conf in pgBackRest Repo Container
With the upgrade to pgBackRest v2.24, it is now necessary to ensure
a configuration setting for pg1-path is available in the pgBackRest
repository host environment when a pgBackRest restore is initiated from
a PG host (e.g. a replica when bootstrapping).  However, being that the
actual value for this setting in the pgBackRest repository host is
irrelevant when performing a restore from a PG host (specifically
because the  setting from the PG host is utilized instead), a dummy
value can be provided that simply ensures a valid directory is set
for pg1-path.  This commit therefore adds a dummy pg1-path setting to
the default pgbackrest.conf file in the pgBackRest repo container, which
ensures a valid pg1-path is configured on the repo host as needed to
support restores from other PG hosts and containers.