Load pgBackRest S3 credentials as secret environmental variables
While the pgBackRest S3 credentials were already stored in secrets,
there was a way the values could be exposed when running Kubernetes
commands, such as "kubectl describe pod" etc.

This moves the pgBackRest S3 credentials to be loaded into a container
using a "secret environmental variable" so the information in the
environmental variable is not revealed even if a user runs a Kubernetes
introspection command.

Issue: [ch7638]
Issue: #1306