Move startup/shutdown to Update & Update User Pwd Var
The primary focus of this commit is to move the startup and shutdown
flags from the scale and scaledown commands associated API endpoints to
the cluster update commands and associated API endpoints, e.g.:

pgo update cluster mycluster --startup
pgo update cluster mycluster --shutdown

When the user specifies that the cluster should be shutdown, a
'Shutdown' boolean parameter the pgcluster spec for the cluster(s)
specified is set to 'false'.  This then triggers the shutdown of the
cluster, i.e. all deployments for the cluster are scaled to 0.
(includes the primary and replica PostgreSQL deployments, as well as
any deployments for supporting services, such as pgBackRest and
pgBouncer). And being that the shutdown itself is triggered by an
update to the pgcluster spec, this means shutdown and startup
functionality is also supported for CR-only workflows.

When the user specifies that the cluster should be started, the
opposite of the above occurs, i.e. all deployments for the cluster are
scaled to 1.  However, it should be noted that the primary is first
started prior to any replicas, ensuring the same node identified as the
primary when the cluster was shutdown is initialized as primary when it
is started back up (and therefore ensuring a consistent architecture
across shutdowns and startups).

In addition to the above, this commit also reverts the new
'PasswordUser' variables and associated '--user-password' CLI flag as
used for cluster creation back to 'Password' and '--password'
respectively.  This ensures backwards compatibility with current
password functionality and associated flags/parameters.