Move rc var Back to Original Location
This commit moves the 'rc' int var and the logic used to populate it
back to its original location.  This var had been moved to an earlier
position in the function when the 'startup/shutdown' functionality
was included with the 'scale/scaledown' commands, but now that
this functionality has been moved to the 'pgo update cluster'
command, this specific code can be moved back down to where
it was originally found.