Add pgo-installer image and required files
These new container images for centos, rhel, and ubi have the needed packages and
files to install the Crunchy PostgreSQL Operator. Each image had the kubectl
client and ansible so that it is able to install the operator from within an
environment. The images are added to the makefile and push scripts where
needed. Each image is built with a pgo-install script that simplifies the
ansible-playbook install command that will be passed in using a kubernetes job.
The permissions on the install script are set to make it an executable file.
The kubernetes repo file is added to the Centos7 image to intall kubectl and the
file is added to the postgres-operator repo under the config directory for the
pgo-installer. The rhel servers ose 3.11 repo is enabled to install kubectl on
the Rhel7 and UBI7 images. Kubectl and oc provide  the same functionality
besides a few additions that oc has. None of these oc specific features are used
by the installer currently so we can use kubectl for both kubernetes and
OpenShift.