Allow tablespaces to be added to existing PostgreSQL clusters
This commit introduces the "pgo update cluster --tablespace"
flag, which allows for users to add additional tablespaces to a
PostgreSQL cluster.

The Operator takes care of creating PVCs for the additional PVCs
for the primary and all replicas in the PostgreSQL cluster, as well
as updating the Deployment template for each instance. This does
have the side effect of causing the PostgreSQL Pods to restart,
as the new volumes need to be mounted.

Issue: [ch7232]