Allow one-off specifications of CPU / Memory
Previously the only way to specify container resources (CPU/RAM)
for a PostgreSQL cluster was to customize the "resource config"
in the pgo.yaml file and reference it from the "pgo create cluster"
command.

This introduces the "--cpu" and "--memory" flags to the
"pgo create cluster" command to allow one to specify custom values
for the PostgreSQL cluster at cluster creation time.

Issue: [ch6658]
Issue: #1268