Updates validation logic on pgBackRest backup options
Currently validation is run on backup options passed in with the --backup-opts
flag to reject unknown or blacklisted flags. This commit updates the validation
to only reject blacklisted flags and allow unknown flags to be passed into the
operator. This will make new pgBackRest features and flags easier to support by
relying on the builtin pgBackRest validation. Now new pgBackrest flags can be
used without updating the list of known flags.

The handleCustomParseErrors function now will take the unknown flag error and
check if the flag is blacklisted. If not, the error is ignored and nil is
returned.

A second level check is added to validate the parsedBackupOpts. Now the return
from handleCustomParseErrors is checked rather than immediately returned.