Postgres Operator automated upgrade status checks
Checks were added for the various PGO cli operations
to validate that the cluster in question is either current to
this release of the Postgres Operator, or has been upgraded to
the current version. This is to ensure that unexpected results
are not achieved due to pgclusters that are out of spec.

A check was also added to ensure the ConfigMap Controller only
syncs configmaps on clusters that have already been upgraded.