Upgrade Postgres Operator script (Bash install)
To ensure that needed resources are not inadvertently
deleted by the user during an upgrade of the Postgres
Operator, this helper script is provided. This script
provides a similar function to the Ansible installation
method's 'update' tag, where the Operator is undeployed,
and the designated namespaces, RBAC rules, pods, etc are
redeployed or recreated as appropriate, but required CRDs
and other resources are left in place.

Note that you will need to update your environment variables
and pgo.yaml configuration files before running this script.