Reorganize API Package for Client Code Generation
This commit reorganizes the content of the v1 package in accordance with
convententions for client code generation.  Most noteably the folder
structure for the Operator API now contains the proper GroupName.  The
deepcopy.go file is also being deleted since the proper deep copy functions
will be generated in subsequent commit.