Updates to Utilize Generated Client Code
This commit updates the PostgreSQL Operator to utilize the client code
generated using the k8s client code generator project.  Therefore, all PGO
clients, as well as all k8s and PGO informers, are no longer created manually,
but rather using this generated code.  This ensures the various informers
and clients are generated according to the latest guidelines and standards,
e.g. those show in the k8s sample controller project. Additionally, the
generated code provides various other benefits, such as mock resources and
interfaces that can be utilized to facilitate unit testing of Operator code.

As part of this change, each a new ControllerManager struct has been created,
which is responsible for creating and managing the group of controllers
needed for each namespace (i.e. controllers to handle events for pgclusters,
pgreplicas, pgtasks, pgpolicies, pods and jobs).  This provides an effective
solution for starting and stopping controllers across all namespaces, and
ensuring controllers are properly cleaned up when a namespace is deleted.
More specifically, contexts and associated cancel functions are generated for
each namespace, which allow the various controllers for a specific namespace
to be easily torn down and removed when a namespace is deleted, and should
therefore no longer be monitored for events.

Additionally, when the various controllers described above are being created
for a namespace, that namespace is now assigned its own unqie clients.  This
mitigates issues seen in certain cloud environments (e.g. GKE) when sharing a
single client across controllers for a large number of namespaces,
specifically throttling/rate-limiting that can occur when a single client
exceeds a certain allowed threshold of requests and/or connections.

And finally, all references to the PostgreSQL Operator 'v1' package reflect
the new directory structure for this package ('apis/crunchydata.com/v1').