Verify & Update Code Gen Scripts
Created update and verify scripts for verifying that generated
code is up-to-date, and then updating that code as needed.  These
scripts are tailored from those included with the Kubernetes sample
controller.  The scripts are stored in a new 'hack' directory, as is
standard in Kubernetes operations for storing various dev related bash
scripts:

https://github.com/kubernetes/kops/blob/master/docs/development/hack.md

Additionally 'verify-codegen' and 'update-codegen' targets have been
added to the Makefile to facilitate running these scripts.  These
targets can be run as part of a CI pipeline to ensure all generated
code remains up-to-date.

And finally, a boilerplate text file is also included to ensure the
proper Crunchy headers are included in all generated code.