Add Code Generation Section to Docs
This commit adds a section to the developer documentation for code
generation.  Specifically, the documentation now explains how code
generation is utilized by the PostgreSQL Operator project, while also
showing how the two new Make targets created for code generation can
be utilized to determine if an update to generated code is required,
and then update that code as needed prior to compiling.