Adds setup steps for pgo client
The `pgo` client needs the client cert and key along with the pgouser file that
contains the admin username and password. This commit updates the
pgo-installer-ansible roles to automatically setup these files for the install.
This allows the user to user the pgo binary locally without having to manually
perform these steps.

The `pgo` binary is automatically downloaded for Linux and MacOS operating
systems.

These automated steps are run on each update and install of the operator when
the `pgo_client_install` varible is true in the inventory file. The steps can
also be run without and update of the operator by using the `setup-client` tag.
This will run the same setup steps against a running instance of the operator.
This tag also requires `pgo_client_install` to be true.