Updates the naming convention of the installers directory
This updates the naming and directory structure of the pgo-installer image and
ansible playbook. The name of the pgo-installer has caused some confusion
considering there are now two sets of ansible playbooks that "install the
operator". The original ansible that runs the install for the operator and now
the new ansible that sets up and runs the kubernetes job for the pgo-installer
image.

Under this directory structure the files needed to build the image are located
under `installers/image` and the pgo-installer ansible playbook is stored under
`installers/method`. The idea behind the `method` directory is that it can be a
home for other installation methods that can be used as a wrapper for the
`pgo-installer` image.