Ensure clone restores a cluster from the new pgBackRest repo
A regression introduced in 1e473a86e would cause the second step of
a cluster clone to fail, as it was attempting to reference the
pgBackRest repository of the source cluster, not the target. As such,
it was unable to find its manifest file and failed.

This switches the reference to be that of the target cluster.

Issue: [ch7748]