Add 'replica' Const & Update Scale Down Msgs
This commit adds a constant for the value of the 'role' label, and
two replica selectors have been updated to use the constant.  Also,
added back the error that is thrown if no replicas are found when
scaling down, which reverting 'replica' back to 'Pgreplica' in the
message returned when scaling down.