Delete Existing rmdata pgtask When Deleting Cluster
When creating a new rmdata pgtask to delete a cluster, an attempt is
first made to delete any pgtask for the cluster with the same name.
This ensures any existing rmdata pgtasks with the same name are first
removed prior to creating the new pgtask, preventing 'pgtask already
exists' errors when attempting to delete a cluster.