remove specDescriptors[].value from CSV
The value field was used to render default values in the OLM UI
but this was not their intent and it prevents installation in 4.3+
openshift clusters. Using alm-examples to imply default hints to
the UI is the best path until CRD/v1 is used and defaulting can
be integrated at the CRD level.