Specify a value for a S3 CA during cluster creation
The PostgreSQL Operator by default ships with the trusted CA for AWS
S3 for ease of integration, though people may use other S3 compatible
interfaces. While one can relatively easily update the primary
"pgo-backrest-repo-config" Secret and set the value of "aws-s3-ca.crt"
to the value of the CA that they wish to use, there may be cases where
one must specify this value on a per-cluster basis.

This introduces the "--pgbackrest-s3-ca-secret" flag on the
"pgo create cluster" command. One must first create a Kubernetes Secret
and put the value of the CA in a key called "aws-s3-ca.crt". The name
of this Secret should be passed as the value to "--pgbackrest-s3-ca-secret",
and the PostgreSQL Operator will use this value as the CA that pgBackRest
will use for pushing to a S3 interface.

Any clones of this cluster will also use this CA for their pgBackRest
S3 interactions.

Issue: [ch6918]
Issue: #1228
Issue: #1337