Use Kuberentes Resource Format in CRD + Resource Specification Cleanup
Prior to this commit, one could specify the resources that the instances
in a PostgreSQL cluster could use by putting in a Kubernetes-format
string into the pgcluster CRD (as well as the pgreplica CRD) in a custom
format that requested both the Kubernetes Request and Limit for Memory
and CPU.

The original impetus for this change came from some research that showed
setting the Limit for Memory would increase the risk that a PostgreSQL
clusters could enter into some less-than-desirable situations where the
OOM killer could take out a PostgreSQL backend and cause every other
connection backend to fail as the PostgreSQL server enters crash
recovery mode. While the Guaranteed QoS in Kubernetes, which requires
both Memory & CPU Limits and Requests to be set, would set an overall
OOM score adjustment to be very low[1] (albeit not the PostgreSQL
recommended score of -1000[2]), this does not affect behavior of
Limit[3] and still puts the PostgreSQL connection backends and risk of
the OOM killer and aforementioned behavior.

As such, this change began with removing the option to set the "Limit"
resources in the pgcluster CRD, which afforded the opportunity to switch
the "Resource" requests to match the the container resource data type
for Kubernetes[4]. However, this was extend to do a few more things to
bring the PostgreSQL Operator inline with best practices for running
PostgreSQL on Kubernetes, including:

- Removing the "ContainerResources" attribute from the pgreplica CRD.
This ensures all failoverable instances have the same resource
definitions. Future work will allow for more customization around this
(though with strong caveats on when to set a replica with mismatched
resources from the primary). You can achieve similar behavior now with
the clone and standby functionality

- Removing the "Default...ContainerResources" attributes from the
PostgreSQL Operator configuration. This provides more flexibility for
how these Pods are deployed and managed by Kubernetes. These include:

- "PostgreSQL containers"
- Load Job
- Backup Jobs
- "rmdata" Jobs
- pgBadger
- pgBouncer

Future work will allow some of these Pods to have their resources set by
the PostgreSQL Operator.

- A small modification to how CPU and Memory are displayed by
`pgo show cluster`, though likely it will go unnoticed

Issue: [ch7766]

[1] https://kubernetes.io/docs/tasks/administer-cluster/out-of-resource/#node-oom-behavior
[2] https://www.postgresql.org/docs/current/kernel-resources.html#LINUX-MEMORY-OVERCOMMIT
[3] https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/#how-pods-with-resource-limits-are-run
[4] https://pkg.go.dev/k8s.io/api/core/v1?tab=doc#ResourceList