Allow CPU / memory resources to be set one-off for pgBouncer Deployments
There are cases where it may be helpful to control the container
resource utilization for the pgBouncer Deployments. As such, this
modifies the pgcluster CRD to store the resource values pertaining to
the pgBouncer Deployments. This also adds new flags to the PostgreSQL
Operator client to manage the CPU / memory resources, including:

- `pgo create cluster --pgbouncer-cpu`
- `pgo create cluster --pgbouncer-memory`
- `pgo create pgbouncer --cpu`
- `pgo create pgbouncer --memory`
- `pgo update pgbouncer --cpu`
- `pgo update pgbouncer --memory`

This also fixes a bug with rotating the pgbouncer system account
password, where all pgBouncer Pods would be checked for propagation
vs. just the pgBouncer Pods for the specific cluster. Also there are
some general code cleanups.