Allow CPU/Memory resources to be set one-off for pgBackRest repositories
There are cases where it may be helpful to control the container
resource utilization for the pgBackRest repository Deployment that comes
with every PostgreSQL cluster that is managed by the PostgreSQL Operato
.

This modifies the pgcluster CRD to store the resource values pertaining
to the pgBackRest repository. This also adds new flags to the PostgreSQL
Operator client to manage the CPU / memory resources, including:

- `pgo create cluster --pgbackrest-cpu`
- `pgo create cluster --pgbackrest-memory`
- `pgo update cluster --pgbackrest-cpu`
- `pgo update cluster --pgbackrest-memory`

This also fixes a bug with rotating the pgbouncer system account
password, where all pgBouncer Pods would be checked for propagation
vs. just the pgBouncer Pods for the specific cluster. Also there are
some general code cleanups.