Remove the "ContainerResources" configuration
This is now replaced with the one-off settings that are available
in the PostgreSQL Operator. It is still possible to achieve similar
behavior by modifying the various Job / Pod templates that are
deployed in the "pgo-config" ConfigMap. However, this helps the
Operator to do things in a more Kubernetes standard way.